requirements.txt
Flask==2.2.2
Werkzeug<3.0.0
ultralytics
torch
torchvision
Pillow
requests
python-dotenv 